+++ apps/fxc-front/src/app/components/ui/about-modal.ts
@@ -59,14 +59,13 @@ export class AboutModal extends LitElement {
             <li>Icons are from <a href="https://icons8.com/" target="_blank">Line Awesome</a></li>
             <li>and many more open source libraries, projects, and artwork</li>
           </ul>
-
-          <h3>
-            <ion-text color="primary"><a href="/privacy-policy">Privacy Policy</a></ion-text>
-          </h3>
-          <h3>
-            <ion-text color="primary"><a href="/terms">Terms and Conditions</a></ion-text>
-          </h3>
         </ion-text>
+        <h3>
+          <ion-text color="primary"><a href="/privacy-policy">Privacy Policy</a></ion-text>
+        </h3>
+        <h3>
+          <ion-text color="primary"><a href="/terms">Terms and Conditions</a></ion-text>
+        </h3>
       </ion-content>
       <ion-footer>
         <ion-toolbar color="light">
