+++ apps/fxc-front/src/app/pages/admin.ts
@@ -4,7 +4,7 @@ import { html, LitElement, PropertyValues, TemplateResult } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { unsafeHTML } from 'lit/directives/unsafe-html.js';
 import { when } from 'lit/directives/when.js';
-import '../ui/google-btn';
+import '../components/ui/google-btn';
 
 const REFRESH_MIN = 10;
