+++ apps/fxc-front/src/app/pages/privacy-policy.ts
