+++ apps/fxc-front/src/app/pages/terms.ts
